#36076 - Moved types - commented debugging logs
