+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "paypal-commercetools-client",
-  "version": "0.1.19-alpha",
+  "version": "0.1.20-alpha",
   "private": false,
   "type": "module",
   "license": "MIT",
